Adding Liv'it v 0.0.1
Liv'it (0.0.1) captures the angst and uncertainty accompanying my
(ongoing) journey from youth to slightly older youth.

I wanted to express how the influence of forces that often promote
established ways of life -- culture, family, society, etc. -- can
pressure us and cause us to be unsure of ourselves. I expand on this
pressure by describing the act of comparing one's current state of life
with what the established norms are for their particular community.
I illustrate the aforementioned uncertainty of oneself by describing
a confused, and admittedly pessimisitic, relationship with hope.