Add Floating 在梦里
This poem was inspired partially by Teresa Tong's song Tián Mì Mì, and also inspired by reading about TPRS (Teaching Proficiency through Reading and Storytelling) on Language Log.

My goal was to convey, in my own way, the quality and type of my dreams lately. I try to describe how immersive my dreams are by indicating that I float in them in a very corporeal sense. Likewise, I detail the types of things I dream about, which are heavily focused around programming. I use a prolonged state of recursion, and abstract tree manipulations, to convey this. Next, I describe my state of mind (and body) when waking from such a dream, which is often in the middle of the night -- my heart is beating quickly because I'm excited about an idea, and for a few precious moments, the abstract world of the dream melds into reality, causing confusion about if I'm still in the state of dream or not.