Add Apart, Together
I wanted to convey the uncertainty of the paths our futures take and the strength of a love so exemplary, that even though the involved parties are separated from one another (maybe by distance, death, or something else), their love has still grown.

I gradually increased the spacing in the lines of each stanza as the poem goes on as a way of emphasizing that the poem continues to be whole and unbroken even as the lines are split further apart from one another.