Add "Edification"
This poem is a product of things that have been on my mind a lot lately, notably climbing (and exercise in general), the notion of salvation espoused by Christianity, and the idea of self-improvement.

I find a great deal of enjoyment in physical exertion. It enables me to sleep better at night, and releases lots of those "feel good" endorphins. But, it also feels purifying to me, in a similar way that shaving my head feels purifying. It doesn't seem a far leap to me to relate the idea of energy being exerted physically by the body with the notion of spiritual and/or moral development. That is the connection I tried to draw in this poem.

Taking things a bit further, if physical exertion can in some way engender spiritual/moral development, then a strong belief in this might be characterized as a kind of "salvation via works" philosophy, which can be (but needn't always be) contrasted with the notion of "salvation via faith".

The program implemented at the bottom illustrates the notion of continual self-improvement. If you find that salvation is wrought with a given amount of muscle tension, sweat, and breathlessness, you'll be disappointed (or happy?) to find that eventually, that same level of muscle tension, sweat, and breathlessness will no longer be enough to maintain your salvation. This is terribly Sisyphean in a sense, but inspiring and humbling when viewed another way.