Add "soften"
This conveys a slightly dramatized version of my internal experience of softening internally since I moved to Charlotte. The moments I've spent by myself, and more importantly, my experiences with others, have made me realize how meaningful authentic connection (with others and oneself) truly is.

I think there are many layers of softening that can apply to a person's life, and I describe some of them in the poem: speech, walking style, breathing, muscular tension, and more abstractly, the notion of the boundaries between (i) oneself and one's ideas of themselves and (ii) oneself and others. I liken a person who hasn't softened to a string instrument that is wound up tightly unbeknownst to themselves -- until, that is, something wiggles those strings a bit loose. In my case, one of these "loosening" experiences came from a remarkable interaction I had with a woman at the Farmer's Market.

This woman had a booth at the market that sold peaches and blackberries, among other things. I saw the peaches and was instantly hooked, but I realized I didn't have cash, which was the only type of payment she accepted. Somewhat disappointed, I started to walk away when, to my surprised the woman asked if I wanted to buy some peaches. I explained that I did, but didn't have any cash to pay with. To my complete surprise, the woman told me not to worry about it: I could take the peaches now, and just pay her later on. I was very confused by this, and I hesitated for a long time trying to figure out if I was in a dream or not. I thought, "What kind of seller at a farmer's market of all places trusts someone to walk off with their hard-won produce?". Still, I eventually came around to the idea and told her that I would bring her the money as soon as I had the opportunity. I took the peaches. A bit later, I found an ATM in a different part of the market that I hadn't noticed previously (this was my first time visiting the market). Anyways, feeling a mix of guilt(?)/the pressure of conscientiousness/the desire to make things right, I immediately withdrew cash from the ATM and walked back to her booth. When the customer who was in front of me moved along, and she saw me again, she gave me the sweetest, most tender and intimate look I've seen from another human being in a long time. Then she said, with an almost inhuman character of gentleness, "You didn't have to do that. I knew you would come back. Sometimes, you have to trust people."

It has been months since then, but I still feel like it was just yesterday when I recall how that last sentence of hers made me feel. It was as if something deep inside me, something non-physical, just...softened. Eased. It was like an internal tightness that I didn't even know about just, well, went away. I felt so light, radiant, and at peace when I left the market. I wish that woman knew how big of an impact her words and actions had on me.