Create faith.md
This poem is about awkwardness, self-judgement, shame, and self-doubt. It reflects times in which I'm at war with myself, and aspiring to disperse a mood or feeling of shame that's arisen in my heart/mind/body.

The language of "a dangerous thing" is deliberately objectifying, reflecting the othering and distancing of one from oneself that occurs when shame is felt. The use of "dangerous" reflects the presence of self-doubt, a lack of trust in the true nature of oneself and one's intentions, and the warring nature of that conflict within the self.

My journey of learning about OCD has resulted in returning often to different thoughts and ideas on shame: why I feel it so acutely and readily, how intrusive thoughts can engender it and reel me in if I'm not careful, and how to alter my relationship to it. A solution, as I understand it, is to offer myself compassion, but sometimes it's also nice to just creatively braindump whatever is real for me at the moment, hence this poem.

I think this poem has an element of tragical yearning to it that I sometimes feel. Recognizing that I so readily offer trust and faith to others, but at times struggle to extend it to myself, is uniquely sad in a way, but also makes me hopeful because awareness is an important step. Noticing the times when I shortchange myself offers me an opportunity to amend that and come into better relationship with myself over time. So, this poem is also about faith and hope in the positive sense as well.