Turned 26 recently, thinking about my mortality and the fragility of life. Pondering loss (I miss you grandma)
Also, the nature and difficulty of personal change at a deep, fundamental level. Of striving to maintain connection and structure in the face of dwindling resources. And, an acknowledgement of finite power, finite time, finite reach. Finding acceptance and grace in that. Finding a warm, compassionate heart to hold myself with in that. And continuing to extend and reach out.