Add 'Everyone In Me Is A Bird'
This poem was a warm up that I made with my writing group earlier tonight. Our inspiration was a line from an Anne Sexton poem called 'In Celebration of My Uterus'. The first line of the poem is 'Everyone in me is a bird.' I took this and ran with it, just letting the words carry me wherever they wanted to.