Add 'A Million Other Things'
This poem is about the creative experience and what it can do and feel like. I really just let the words flow on this one; I didn't try to constrain myself or make it make too much sense. I felt something pouring out of me, resonating in a nice corresponding way, as I wrote this. I tried to ride that feeling and give it sway on the page, and then transferred it here.