Add full-gutted.md
I was thinking about A.E. Housman's poem, "Be still my soul, be still": https://poets.org/poem/be-still-my-soul-be-still-arms-you-bear-are-brittle

I have been considering David Whyte's book, "The Heart Aroused: Poetry and the Preservation of the Soul in Corporate America", and thinking about the discussions one of the chapters had about our awareness of gut sensations. I advance the notion of a "full-gutted" approach to life being an open exposure to the vulnerability, risk, and uncertainty entailed in our very existence, with all the soul-crushing and incomparably joyful moments that entails.

I imagine myself and others feeling our stomach drop in much the way we do when a plane is landing, or we're skydiving, or we're riding a rollercoaster. All those sensations can be a lot for the body to tune into, but increasing I feel my life demands that presence to what is truly there.